requirements.txt
lxml==4.6.3
python-docx==0.8.10
odfpy==1.4.1
